Fixed bug.
np.argsort(x) does not work correctly when x is a pandas object that contains NaN values. The results contained -1 indexes for the NaN values, and this in turn caused ValueError later on. If we apply the argsort function to a list and not directly to the pandas object, it works correctly.