Changed how merge works.
All data in the list of files ('filenames') will always be combined/merged with themselves. Keyword 'merge' does not affect this.
If 'merge' is False, that's the only merging that happens.
If 'merge' is True, then this data will also be merged with any pre-existing data in tplot.
This is a different logic than what we had before in this function. This logic should also be the same in cdf_to_tplot.