Fixed bug, download.py will now reject responses from captive portals and will not replace existing cdf and netcdf files.
In some cases with captive wifi portals, the wifi responds with 200-OK but the downloaded file is not the file requested by the user, it only contains some html code, usually asking the user to login into the system. Previously, this fake file could replace an already existing cdf file, and then it would be deleted since it cannot be loaded into tplot. This bug was fixed, and now download.py will reject the fake file before replacing any existing files. This only works for cdf and netcdf files.