fix an error message when user=None
I got an error on the string concatenation
  "'Invalid password for user: ' + user + [...]

because in my case, user was `None`. I could not reproduce this error later,
I believe that's because `user=None` was saved in `~/mms_auth_info.pkl` by
a previous version of pyspedas, while after deleting this file, it was
recreated storing `user=''`.

Nevertheless, someone else may hit this problem and the fix is easy enough,
explicitly converting `user` to a string.